fixed variable env tests,  moved example envs to pettingzoo repo proper so that downstream packages can use them (i.e. supersuit)
