+++ .github/workflows/linux-test.yml
@@ -29,6 +29,7 @@ jobs:
       run: |
         pytest ./test/pytest_runner.py
         pytest ./test/all_parameter_combs.py
+        pytest ./test/variable_env_test.py
     - name: Bandit, Codespell, and Flake 8 tests
       run: |
         bash ./check_style.sh
